+++ src/helpers/getRecommentedFood.js
@@ -8,3 +8,7 @@ export const getRecommentedFood = async () => {
     console.log(err);
   }
 };
+
+
+
+
